git-svn-id: https://plugins.svn.wordpress.org/get-better-excerpt/trunk@124762 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ get-better-excerpt.php
@@ -6,7 +6,7 @@ Description: An easy to use WordPress function to add scheduled posts to any the
 Author: Christopher Ross
 Tags: future, upcoming posts, upcoming post, upcoming, draft, Post, scheduled, preview
 Author URI: http://thisismyurl.com
-Version: 0.1.2
+Version: 0.1.3
 */
 
 /*  Copyright 2008  Christopher Ross  (email : info@thisismyurl.com)
@@ -63,33 +63,38 @@ function Easyget_better_excerpt_options() {
  <?php  
 	if (function_exists(zip_open)) {
 	$file = "get-better-excerpt";
-	$pluginUpdate = file_get_contents('http://downloads.wordpress.org/plugin/'.$file.'.zip');
-	$myFile = "../wp-content/uploads/cache-".$file.".zip";
-	$fh = fopen($myFile, 'w') or die("can't open file");
-	$stringData = $pluginUpdate;
-	fwrite($fh, $stringData);
-	fclose($fh);
-	
-	$zip = zip_open($myFile);
-	while ($zip_entry = zip_read($zip)) {
-		if (zip_entry_name($zip_entry) == $file."/".$file.".php") {$size = zip_entry_filesize($zip_entry);}
-	}
-	zip_close($zip);
-	unlink($myFile);
-	
-	if ($size != filesize("../wp-content/plugins/".$file."/".$file.".php")) {?>    
-    <div id="sm_pnres" class="postbox">
-        <h3 class="hndle"><span>Plugin Status</span></h3>
-        <div class="inside">
-        <ul class='options'>
-        <style>.options a {text-decoration:none;}</style>
-        <li>This plugin is out of date. <a href='http://downloads.wordpress.org/plugin/<?php echo $file;?>.zip'>Please <strong>download</strong> the latest version.</a></li>
-        </ul>
-        </div>
-        </div>
-    <?php 
+			$lastupdate = get_option($file."-update");
+		if (strlen($lastupdate )==0 || date("U")-$lastupdate > $lastupdate) {
+			$pluginUpdate = file_get_contents('http://downloads.wordpress.org/plugin/'.$file.'.zip');
+			$myFile = "../wp-content/uploads/cache-".$file.".zip";
+			$fh = fopen($myFile, 'w') or die("can't open file");
+			$stringData = $pluginUpdate;
+			fwrite($fh, $stringData);
+			fclose($fh);
+			
+			$zip = zip_open($myFile);
+			while ($zip_entry = zip_read($zip)) {
+				if (zip_entry_name($zip_entry) == $file."/".$file.".php") {$size = zip_entry_filesize($zip_entry);}
+			}
+			zip_close($zip);
+			unlink($myFile);
+			
+			if ($size != filesize("../wp-content/plugins/".$file."/".$file.".php")) {?>    
+			<div id="sm_pnres" class="postbox">
+				<h3 class="hndle"><span>Plugin Status</span></h3>
+				<div class="inside">
+				<ul class='options'>
+				<style>.options a {text-decoration:none;}</style>
+				<li>This plugin is out of date. <a href='http://downloads.wordpress.org/plugin/<?php echo $file;?>.zip'>Please <strong>download</strong> the latest version.</a></li>
+				</ul>
+				</div>
+				</div>
+	<?php
+		} 
+		update_option($file."-update", date('U'));
     }}
 	?>
+
     </div>
     </div>
